Rename files and variables
