+++ src/apps/controllers/HomeController.js
@@ -1,5 +1,3 @@
-class HomeController {
-    
-  }
-  
-  module.exports = new HomeController();
+class HomeController {}
+
+module.exports = new HomeController();
