--- src/apps/controllers/RecommedController.js
@@ -1,3 +1,0 @@
-class RecommedController {}
-
-module.exports = new RecommedController();
