+++ src/apps/controllers/RecommendController.js
@@ -1,0 +1,3 @@
+class RecommendController {}
+
+module.exports = new RecommendController();
