+++ src/apps/controllers/index.js
@@ -1,13 +1,13 @@
 const AuthController = require("../controllers/AuthController");
 const HomeController = require("../controllers/HomeController");
 const CombinationController = require("../controllers/CombinationController");
-const RecommedController = require("../controllers/RecommedController");
-const WelcomController = require("./WellcomController");
+const RecommendController = require("../controllers/RecommendController");
+const WelcomeController = require("../controllers/WelcomeController");
 
 module.exports = {
   AuthController,
   HomeController,
   CombinationController,
-  RecommedController,
-  WelcomController
+  RecommendController,
+  WelcomeController
 };
