+++ src/config/database/index.js
@@ -1,0 +1,1 @@
+
