+++ src/routes/combinationRoute.js
@@ -2,7 +2,4 @@ const express = require("express");
 const router = express.Router();
 const CombinationController = require("../apps/controllers/CombinationController");
 
-
-
-
 module.exports = router;
