+++ src/routes/index.js
@@ -1,14 +1,14 @@
 const homeRouter = require("./homeRoute");
-const welcomeRouter = require("./wellcomeRoute");
+const welcomeRouter = require("./welcomeRoute");
 const authRouter = require("./authRoute");
 const combinationRouter = require("./combinationRoute");
-const recommedRouter = require("./recommedRoute");
+const recommendRouter = require("./recommendRoute");
 
 function routes(app) {
   app.use("/home", homeRouter);
   app.use("/auth", authRouter);
   app.use("/combination", combinationRouter);
-  app.use("/recommed", recommedRouter);
+  app.use("/recommend", recommendRouter);
   app.use("/", welcomeRouter);
 }
