--- src/routes/recommedRoute.js
@@ -1,5 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const RecommedController = require("../apps/controllers/RecommedController");
-
-module.exports = router;
